added km to mi conversion and moved to conversions.swift
