+++ ConvertRunDistances/ViewController.swift
@@ -9,23 +9,30 @@
 import UIKit
 
 class ViewController: UIViewController {
-    
+//    let conversionConstant = 1.6129
+
     var miles = 0.0
     var kilometers = 0.0
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        convertMilesToKilometers(miles: 26.2)
+        convertMilesToKilometers(miles: 50.0)
+        convertKilometersToMiles(kilometers: 100.0)
     }
     
-    func convertMilesToKilometers(miles: Double) {
-        let conversionConstant = 1.6129
-        print(miles)
-        print(conversionConstant)
-        
-        let computedKilometers = conversionConstant * (miles)
-        print("\(miles) mi is \(computedKilometers) km ")
-    }
+//    func convertMilesToKilometers(miles: Double) {
+//        print(miles)
+//        
+//        let computedKilometers = conversionConstant * (miles)
+//        print("\(miles) mi is \(computedKilometers) km ")
+//    }
+    
+//    func convertKilometersToMiles(kilometers: Double) {
+//        print(kilometers)
+//
+//        let computedMiles = (kilometers) / conversionConstant
+//        print("\(kilometers) km is \(computedMiles) km ")
+//    }
 }
